requirements.txt
psycopg[binary]==3.3.3
python-dotenv==1.0.1
pandas==2.2.2
tqdm==4.66.4
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
